# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAMAZIRT API Documentation",
        "description": "API for TAMAZIRT",
        "contact": {
            "email": "support@tamazirt.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "TAMAZIRT API Server"
        }
    ],
    "paths": {
...
        "/api/tourist/navigators": {
            "get": {
                "tags": [
                    "Tourist Actions"
                ],
                "summary": "Get verified navigators (Public)",
                "description": "Retrieves a paginated list of verified navigator profiles. Can be filtered by city and language.",
                "operationId": "getVerifiedNavigators",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter navigators by city.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Fez"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Filter navigators by language spoken.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "English"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Paginated list of navigator profiles.",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
        "/api/tourist/requests": {
            "get": {
                "tags": [
                    "Tourist Actions"
                ],
                "summary": "Get own service requests (Tourist only)",
                "description": "Retrieves a list of all service requests created by the authenticated tourist. Requires 'tourist' role.",
                "operationId": "touristMyRequests",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Pending"
                                                    },
                                                    "navigator": {
                                                        "description": "Associated navigator details",
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: tourist access required."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tourist Actions"
                ],
                "summary": "Send a service request (Tourist only)",
                "description": "Allows an authenticated tourist to send a new service request to a specific navigator. Requires 'tourist' role.",
                "operationId": "touristStoreRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "navigator_id",
                                    "date",
                                    "destination",
                                    "number_of_people"
                                ],
                                "properties": {
                                    "navigator_id": {
                                        "description": "ID of the target navigator.",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date": {
                                        "description": "Date and time of the request (must be in the future).",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-01T10:00:00"
                                    },
                                    "destination": {
                                        "description": "The requested destination.",
                                        "type": "string",
                                        "example": "Old Medina, Fez"
                                    },
                                    "number_of_people": {
                                        "description": "Number of people in the group (min 1).",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "We are looking for a 3-hour tour in the morning.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request sent successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request sent successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: tourist access required."
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tourist/requests/{id}": {
            "delete": {
                "tags": [
                    "Tourist Actions"
                ],
                "summary": "Cancel/Delete own service request (Tourist only)",
                "description": "Allows the authenticated tourist to delete/cancel one of their service requests. Requires ownership.",
                "operationId": "touristDeleteRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the request to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: not the request owner."
                    },
                    "404": {
                        "description": "Request not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tourist Actions"
                ],
                "summary": "Update own service request (Tourist only)",
                "description": "Allows the authenticated tourist to update an existing request, but only if its status is 'Pending'. Requires ownership.",
                "operationId": "touristUpdateRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the request to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "destination",
                                    "number_of_people"
                                ],
                                "properties": {
                                    "date": {
                                        "description": "Updated date and time.",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-05T14:00:00"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "New City Center"
                                    },
                                    "number_of_people": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "New message for the navigator.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Request updated successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Cannot edit a non-Pending request."
                    },
                    "403": {
                        "description": "Unauthorized: not the request owner."
                    },
                    "404": {
                        "description": "Request not found."
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tourist/profile": {
            "put": {
                "tags": [
                    "Tourist Actions"
                ],
                "summary": "Update own user profile (Tourist only)",
                "description": "Allows the authenticated tourist to update their name, email, and profile picture. Requires 'tourist' role.",
                "operationId": "touristUpdateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Tourist Name",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "new.tourist@example.com",
                                        "nullable": true
                                    },
                                    "avatar": {
                                        "description": "New profile image file (max 5MB).",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully!"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: tourist access required."
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer Token Authentication",
                "in": "header",
                "bearerFormat": "Sanctum Token",
                "scheme": "Bearer"
            }
        }
    },
...
}